SAT-804 (#363)
* stake fix

* stake fix

* upd mock

+++ lib/solana/interface.go
@@ -55,7 +55,7 @@ type Interface interface {
 	GetNFTMintAddrs(ctx context.Context, walletAddr string) ([]string, error)
 	GetNFTMetadata(mintAddr string) (*ArweaveNFTMetadata, error)
 	InitializeStakePool(ctx context.Context, feePayer, issuer types.Account, asset common.PublicKey) (txHast string, stakePool types.Account, err error)
-	Stake(ctx context.Context, feePayer, userWallet types.Account, pool, asset common.PublicKey, duration int64, amount uint64) (string, error)
+	Stake(ctx context.Context, feePayer, userWallet types.Account, pool, asset common.PublicKey, duration int64, amount float64) (string, error)
 	Unstake(ctx context.Context, feePayer, userWallet types.Account, stakePool, asset common.PublicKey) (string, error)
 }
